refactor(kv): extract `entity_type` module to separate file
